refactor: place code from bff to app module
