+++ bff-service/src/app.module.ts
@@ -2,12 +2,10 @@ import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { BffModule } from './bff/bff.module';
-import { BffController } from './bff/bff.controller';
 
 @Module({
-  imports: [ConfigModule.forRoot(), BffModule],
-  controllers: [AppController, BffController],
+  imports: [ConfigModule.forRoot()],
+  controllers: [AppController],
   providers: [AppService],
 })
 export class AppModule { }
