--- bff-service/src/bff/bff.module.ts
@@ -1,5 +1,0 @@
-import { Module } from '@nestjs/common';
-import { BffController } from './bff.controller';
-
-@Module({ controllers: [BffController] })
-export class BffModule { }
